% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_cycles.R
\name{add_dummy_day}
\alias{add_dummy_day}
\title{Add Dummy Day 21}
\usage{
add_dummy_day(df)
}
\arguments{
\item{df}{Processed cycle data}
}
\value{
Data frame with day 21 dummy observations added
}
\description{
Add day 21 as a dummy variable representing day -14 for cyclical continuity.
}
