% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{create_fourier_basis}
\alias{create_fourier_basis}
\title{Create Fourier Basis Matrix}
\usage{
create_fourier_basis(days, n_harmonics, cycle_length)
}
\arguments{
\item{days}{Vector of cycle days}

\item{n_harmonics}{Number of Fourier harmonics}

\item{cycle_length}{Expected cycle length in days}
}
\value{
Design matrix with intercept and sin/cos terms
}
\description{
Create Fourier basis matrix for cyclic regression.
}
\keyword{internal}
