% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploratory.R
\name{data_summary}
\alias{data_summary}
\title{Data Summary}
\usage{
data_summary(
  period_dates,
  outcome_data,
  confounder_data = NULL,
  id_col = "user_id",
  date_col = "period_date",
  outcome_col = "outcome",
  outcome_date_col = "date"
)
}
\arguments{
\item{period_dates}{Period dates data}

\item{outcome_data}{Outcome data}

\item{confounder_data}{Confounder data (optional)}

\item{id_col}{User ID column name}

\item{date_col}{Period date column name}

\item{outcome_col}{Outcome column name}

\item{outcome_date_col}{Outcome date column name}
}
\value{
Data frame with summary statistics
}
\description{
Generate a summary table of the input data.
}
