% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{find_turning_points}
\alias{find_turning_points}
\title{Find Turning Points}
\usage{
find_turning_points(
  gam_result,
  find_peaks_troughs = TRUE,
  round_to_whole_days = TRUE,
  day_range = c(-14, 13)
)
}
\arguments{
\item{gam_result}{Result from fit_gam()}

\item{find_peaks_troughs}{If TRUE, find turning points (first derivative sign change).
If FALSE, find points where curvature changes sign. Default TRUE.}

\item{round_to_whole_days}{If TRUE, round results to nearest whole day. Default TRUE.}

\item{day_range}{Only return points within this range. Default c(-14, 13).}
}
\value{
Vector of cycle days where turning points occur
}
\description{
Find turning points (peaks/troughs) of the GAM curve.
}
