% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{mc_plot_confounders}
\alias{mc_plot_confounders}
\title{Plot Confounder Effects}
\usage{
mc_plot_confounders(results, title = "Confounder Effects", ...)
}
\arguments{
\item{results}{mcanalysis object from mc_analyze()}

\item{title}{Plot title}

\item{...}{Additional arguments}
}
\value{
ggplot object
}
\description{
Plot Confounder Effects
}
