% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{mc_summary}
\alias{mc_summary}
\title{Print Summary of Menstrual Cycle Analysis}
\usage{
mc_summary(results)
}
\arguments{
\item{results}{mcanalysis object from mc_analyze()}
}
\description{
Print Summary of Menstrual Cycle Analysis
}
