% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{normalize_outcome}
\alias{normalize_outcome}
\title{Normalize Outcome}
\usage{
normalize_outcome(df, outcome_col = "outcome")
}
\arguments{
\item{df}{Dataset with outcome variable}

\item{outcome_col}{Name of outcome column (default "outcome")}
}
\value{
Data frame with normalized outcome 'a' added
}
\description{
Normalize outcome to percentage of each user's mean.
}
