% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{plot.mcanalysis}
\alias{plot.mcanalysis}
\title{Plot MC Analysis Results}
\usage{
\method{plot}{mcanalysis}(x, ...)
}
\arguments{
\item{x}{mcanalysis object}

\item{...}{Additional arguments passed to plot_cycle_effect()}
}
\description{
Plot MC Analysis Results
}
