% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploratory.R
\name{plot_confounder_distributions}
\alias{plot_confounder_distributions}
\title{Plot Confounder Distributions}
\usage{
plot_confounder_distributions(
  confounder_data,
  id_col = "user_id",
  confounder_cols = NULL,
  title = "Confounder Distributions",
  color = "mediumpurple"
)
}
\arguments{
\item{confounder_data}{DataFrame with confounder variables}

\item{id_col}{Name of user ID column (excluded from plots)}

\item{confounder_cols}{Vector of confounder column names. If NULL, uses all numeric columns.}

\item{title}{Plot title}

\item{color}{Histogram fill color}
}
\value{
ggplot2 object
}
\description{
Plot distributions of confounder variables.
}
