% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_confounder_effects}
\alias{plot_confounder_effects}
\title{Plot Confounder Effects}
\usage{
plot_confounder_effects(
  confounder_results,
  title = "Confounder Effects on Cycle-Outcome Relationship"
)
}
\arguments{
\item{confounder_results}{Results from analyze_confounders()}

\item{title}{Plot title}
}
\value{
ggplot2 object
}
\description{
Plot forest plot of confounder effects.
}
