% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploratory.R
\name{plot_cycle_lengths}
\alias{plot_cycle_lengths}
\title{Plot Cycle Length Distribution}
\usage{
plot_cycle_lengths(
  period_dates,
  id_col = "user_id",
  date_col = "period_date",
  title = "Cycle Length Distribution",
  color = "coral",
  valid_range = c(21, 35)
)
}
\arguments{
\item{period_dates}{DataFrame with period start dates}

\item{id_col}{Name of user ID column}

\item{date_col}{Name of date column}

\item{title}{Plot title}

\item{color}{Histogram fill color for valid cycles}

\item{valid_range}{Vector of c(min, max) valid cycle length}
}
\value{
List with ggplot2 object and summary data frame
}
\description{
Plot distribution of menstrual cycle lengths.
}
