% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploratory.R
\name{plot_observations_per_user}
\alias{plot_observations_per_user}
\title{Plot Observations per User}
\usage{
plot_observations_per_user(
  outcome_data,
  id_col = "user_id",
  title = "Observations per User",
  color = "#008080"
)
}
\arguments{
\item{outcome_data}{DataFrame with outcome data}

\item{id_col}{Name of user ID column}

\item{title}{Plot title}

\item{color}{Histogram fill color}
}
\value{
List with ggplot2 object and summary data frame
}
\description{
Plot distribution of number of observations per user.
}
