% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploratory.R
\name{plot_outcome_distribution}
\alias{plot_outcome_distribution}
\title{Exploratory Data Analysis Functions}
\usage{
plot_outcome_distribution(
  outcome_data,
  outcome_col = "outcome",
  title = "Outcome Distribution",
  color = "steelblue",
  bins = 50
)
}
\arguments{
\item{outcome_data}{DataFrame with outcome values}

\item{outcome_col}{Name of outcome column}

\item{title}{Plot title}

\item{color}{Histogram fill color}

\item{bins}{Number of histogram bins}
}
\value{
ggplot2 object
}
\description{
Functions for exploring menstrual cycle data before analysis.
Plot Outcome Distribution
}
\details{
Plot histogram and boxplot of the outcome variable.
}
