% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{print.mcanalysis}
\alias{print.mcanalysis}
\title{Print MC Analysis Results}
\usage{
\method{print}{mcanalysis}(x, ...)
}
\arguments{
\item{x}{mcanalysis object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print MC Analysis Results
}
